Add from argument to allow pagination
Strategy choosen is to return cases with an id GREATER THAN the given
id in the 'from' argument. This allows the first request to leave out
the 'from' argument, then on subsequent calls, use the last returned
id from the previous request as the 'from' argument value..

Otherwise, we need to add a 'more' and 'next' fields to the returned
list, these would be expected to only exist if there is more data, so
we must do some lookahead on the table to avoid confusion on the client
side. With the choosen strategy, we would get an extra request, but the
client is prepared to receive a 404 when there is no more cases.

Auditors: jbn, atm, jbr