Add state transition information
to clarify when a case can be payed and which tasks is payable

This lead to the addition of the PENDING_CLOSE state, used by the
editors when a reviewer rejects a case due to a bkm assessment
resolving to 'not relevant'. The bkm task must be payed before the case
can be moved to the CLOSED state.

Auditors: jbn, atm, jbr