Initial skeleton for payment preview
The execute() function (pro2-161) should return csv. It is
most likely that the preview() function should instead return the same
data but as json for the gui to format and display..
This has to be decided upon.

Meanwhile, while doing the first work on payment procedures, this is
handy since we need to compare with existing payment sheets.

Auditors: jbn, atm, jbr