Apply changes to deliver information needed by the gui
The gui needs to show, not only the payments, but also which reviewers
will receive the payment, and for which faustnumbers is the payments
created.

There is also a need for sorting the payments so that payments for a specific user
or faustnumber can quickly be identified. Allthough it can be done by the gui,
I find it better to handle this in the backend to guarantee that the gui never
modifies payment information, which could lead to confusion or erroneous case modifications

Auditors: jbn, atm, jbr