Handle changing status to PENDING_CLOSE
When a bkm assessment is requested, and returned with 'not relevant'.
The case is placed in status PENDING_CLOSE and the BKM assessment task
is approved (done by the backend).

When cases is payed, the BKM task will be payed, then the case is
moved to status CLOSED

Auditors: jbn, atm, jbr