Use new version of openformat connector
which fixes a problem with two properties that may or may not be an array.

Rewrite a few parts of the caseupdater to a more safe and readable state.

Use ninja powers to add the 'metakompassubjects' property, also returned
by openformat (will be used by later update stories to approve METAKOMPAS
tasks automatically)

Auditors: jbn, atm, jbr