Change filtering for invalid cases when exporting to dataio
since only the BRIEF task has a faustnumber assigned.

Also check for cases without a BRIEF task since this is mandatory,
and in general check that there is tasks on the case, so that we
know that cases exported is valid before the reach the conversion flow script.

This adds a few extra cases hence the payment test needs to be modified.
This has no relevance to the change - just a sideeffect

Auditors: jbn, atm, jbr, mib