Normalize catalog codes before inserting them into codes
to ensure that we can always search them using uppercasing.
(this is the ususal form, but there may be stray codes existing in
the records in lower case - and editors may use lowercase when searching).

We could solve this in the JsonbContains... function, but this would affect
other future uses of this - not necessarily immune towards casing.

Auditors: jbn, atm, jbr, mib