Add initial message to a case depending on casenote AND tasks
Since currently, the GUI does not provide the casenote - so
we need to determine a proper message from the list of tasks.

We currently have no standard phrases/texts defined globally, so
the text is hardcoded - as it is in the notification templates -
but we hope to add this in the comming years.

Also check if there is any text to display. If the message text
is empty, there is no need to create an initial message

Auditors: jbn, atm, jbr, mib