Add testdata for all fields that is to be anonymized
just to be absolutely sure they are cleared.

Also make it possible to re-activate a previously deactivated/anonymized
user by clearing the deactivated field when active status changes to 'active'.
(otherwise, the only way to reactivate users would be by doing it directly
in the db or create a new user with the same name, which is kind of messy)

Auditors: jbn, atm, jbr, mib