Rework use of assigned field
by not allowing it to be set from the outside (it has always been an internal
field and should never have been exposed for external manipulation)

Then update the assigned field every time a new reviewer is assigned.

(the assigned field will be shown in the GUI as 'oprettet' hence afterwards.

Auditors: atm, jbr, mib